backend/requirements.txt
Flask
flask_cors 
bcrypt 
python-dotenv
PyJWT
psycopg2

